Phase 3a (part 2): convert db.py into a db/ package
Split the single 699-line db.py into a package so the access layer can grow
(typed query helpers, Phase 3b migrations) without becoming a god-module:

- db/connection.py — get_connection + db_connection context manager
- db/schema.py     — every CREATE TABLE statement, grouped by source
- db/github.py     — typed query helpers for the github_* tables
- db/__init__.py   — re-exports the full public API

db/__init__.py re-exports all 11 previously-importable symbols, so every
`from rebalance.ingest.db import ...` site (40+ across src/tests/scripts)
keeps working with zero changes.

Behavior-preserving: full suite identical to baseline (285 passed; same one
pre-existing failure and one pre-existing collection error).
