+++ src/rebalance/ingest/db/schema.py
@@ -1,84 +1,14 @@
-"""
-Shared database layer for rebalance — connection factory, schema creation,
-sqlite-vec extension loading, and context managers.
+"""All CREATE TABLE statements for rebalance, grouped by source.
 
-All CREATE TABLE statements live here so the full DB shape is visible in
-one place.  Individual modules call the appropriate ensure_*_schema()
-function (or use the db_connection context manager) rather than carrying
+Keeping every DDL statement in one module means the full DB shape is visible
+in one place; individual modules call the appropriate ``ensure_*_schema()``
+function (or use the ``db_connection`` context manager) rather than carrying
 their own DDL.
 """
 
 from __future__ import annotations
 
 import sqlite3
-from contextlib import contextmanager
-from pathlib import Path
-from typing import Callable, Generator
-
-try:
-    import sqlite_vec
-except Exception:  # pragma: no cover - import guard for environments without sqlite-vec
-    sqlite_vec = None
-
-
-# ---------------------------------------------------------------------------
-# Connection factory
-# ---------------------------------------------------------------------------
-
-
-def get_connection(database_path: Path) -> sqlite3.Connection:
-    """Open a SQLite connection with WAL mode, foreign keys, and sqlite-vec loaded.
-
-    Note: sqlite-vec may not load on all Python builds (e.g., system Python without
-    extension support). The connection will still work for basic queries.
-    """
-    database_path.parent.mkdir(parents=True, exist_ok=True)
-    conn = sqlite3.connect(str(database_path))
-    conn.execute("PRAGMA journal_mode=WAL")
-    conn.execute("PRAGMA foreign_keys=ON")
-    # Wait up to 30s for a writer slot before raising "database is locked".
-    # Background refreshes from the TUI can briefly overlap launchd jobs;
-    # without this they fail instantly. See git history for the 2026-05 incident.
-    conn.execute("PRAGMA busy_timeout=30000")
-
-    # Try to load sqlite-vec, but gracefully fall back if unavailable
-    try:
-        if sqlite_vec is not None and hasattr(conn, 'enable_load_extension'):
-            conn.enable_load_extension(True)
-            sqlite_vec.load(conn)
-            conn.enable_load_extension(False)
-    except (AttributeError, Exception):
-        # sqlite-vec not available on this Python build; continue without it
-        pass
-
-    conn.row_factory = sqlite3.Row
-    return conn
-
-
-@contextmanager
-def db_connection(
-    database_path: Path,
-    ensure_fn: Callable[[sqlite3.Connection], None] | None = None,
-) -> Generator[sqlite3.Connection, None, None]:
-    """Context-managed database connection with optional schema setup.
-
-    Usage::
-
-        with db_connection(db_path, ensure_schema) as conn:
-            rows = conn.execute("SELECT ...").fetchall()
-
-    The connection is always closed on exit — even if the caller raises.
-    Pass *ensure_fn* to guarantee a specific set of tables exists (e.g.
-    ``ensure_schema``, ``ensure_calendar_schema``).  Omit it when you only
-    need a bare connection.
-    """
-    conn = get_connection(database_path)
-    if ensure_fn is not None:
-        ensure_fn(conn)
-    try:
-        yield conn
-    finally:
-        conn.close()
 
 
 # ---------------------------------------------------------------------------
@@ -605,74 +535,6 @@ def ensure_github_schema(conn: sqlite3.Connection) -> None:
     _ensure_github_artifact_schema(conn)
     _ensure_github_knowledge_schema(conn)
     conn.commit()
-
-
-# ---------------------------------------------------------------------------
-# GitHub activity queries (read helpers — keep DDL above, DML below)
-# ---------------------------------------------------------------------------
-
-
-def top_active_repos(
-    conn: sqlite3.Connection, top_n: int, since_days: int = 7
-) -> list[str]:
-    """Repo full-names ranked by recent activity score, highest first.
-
-    Score sums ``commits + prs_opened + prs_merged + issues_opened +
-    issue_comments + reviews`` over the trailing *since_days* window
-    (``github_activity.scan_date``). Repos with a zero score are excluded.
-    """
-    rows = conn.execute(
-        """
-        SELECT repo_full_name,
-               SUM(commits) + SUM(prs_opened) + SUM(prs_merged) +
-               SUM(issues_opened) + SUM(issue_comments) + SUM(reviews) AS score
-        FROM github_activity
-        WHERE scan_date >= date('now', ?)
-        GROUP BY repo_full_name
-        HAVING score > 0
-        ORDER BY score DESC
-        LIMIT ?
-        """,
-        (f"-{since_days} days", top_n),
-    ).fetchall()
-    return [r[0] for r in rows]
-
-
-def repo_last_active(
-    conn: sqlite3.Connection, repos: list[str] | None = None
-) -> dict[str, str]:
-    """Map of ``repo_full_name`` -> latest ``last_active_at`` (raw ISO string).
-
-    Pass *repos* to restrict to a subset; omit it for every repo. Repos with
-    no non-null ``last_active_at`` are omitted. Timestamp parsing is left to
-    the caller — this helper only touches the database.
-    """
-    if repos is not None:
-        if not repos:
-            return {}
-        placeholders = ",".join("?" * len(repos))
-        rows = conn.execute(
-            f"SELECT repo_full_name, MAX(last_active_at) FROM github_activity "
-            f"WHERE repo_full_name IN ({placeholders}) "
-            f"AND last_active_at IS NOT NULL "
-            f"GROUP BY repo_full_name",
-            list(repos),
-        ).fetchall()
-    else:
-        rows = conn.execute(
-            "SELECT repo_full_name, MAX(last_active_at) FROM github_activity "
-            "WHERE last_active_at IS NOT NULL GROUP BY repo_full_name"
-        ).fetchall()
-    return {r[0]: r[1] for r in rows}
-
-
-def repo_meta_names(conn: sqlite3.Connection) -> set[str]:
-    """Set of ``repo_full_name`` values present in ``github_repo_meta``.
-
-    This is the canonical "fully synced" watched set — a repo only gets a
-    ``github_repo_meta`` row after a complete artifact sync.
-    """
-    return {r[0] for r in conn.execute("SELECT repo_full_name FROM github_repo_meta")}
 
 
 # ---------------------------------------------------------------------------
